--- a/tags.go
+++ b/tags.go
@@ -1,20 +1,16 @@
 package stats
 
 import (
-	"bytes"
-	"fmt"
-	"regexp"
 	"sort"
-)
-
-const (
-	tagPrefix   = ".__"
-	tagSep      = "="
-	tagFailsafe = "_"
+	"strings"
 )
 
 // illegalTagValueChars are loosely set to ensure we don't have statsd parse errors.
-var illegalTagValueChars = regexp.MustCompile(`[:|.]`)
+var illegalTagValueCharsReplacer = strings.NewReplacer(
+	".", "_",
+	":", "_",
+	"|", "_",
+)
 
 type tagPair struct {
 	dimension string
@@ -28,19 +24,30 @@ func (t tagSet) Swap(i, j int)      { t[i], t[j] = t[j], t[i] }
 func (t tagSet) Less(i, j int) bool { return t[i].dimension < t[j].dimension }
 
 func serializeTags(tags map[string]string) string {
+	const prefix = ".__"
+	const sep = "="
+
 	if len(tags) == 0 {
 		return ""
 	}
-	tagPairs := make([]tagPair, 0, len(tags))
-	for tagKey, tagValue := range tags {
-		tagValue = illegalTagValueChars.ReplaceAllLiteralString(tagValue, tagFailsafe)
-		tagPairs = append(tagPairs, tagPair{tagKey, tagValue})
+	pairs := make([]tagPair, 0, len(tags))
+	n := (len(prefix) + len(sep)) * len(tags)
+	for k, v := range tags {
+		n += len(k) + len(v)
+		pairs = append(pairs, tagPair{
+			dimension: k,
+			value:     illegalTagValueCharsReplacer.Replace(v),
+		})
 	}
-	sort.Sort(tagSet(tagPairs))
+	sort.Sort(tagSet(pairs))
 
-	buf := new(bytes.Buffer)
-	for _, tag := range tagPairs {
-		fmt.Fprint(buf, tagPrefix, tag.dimension, tagSep, tag.value)
+	var w strings.Builder
+	w.Grow(n)
+	for _, tag := range pairs {
+		w.WriteString(prefix)
+		w.WriteString(tag.dimension)
+		w.WriteString(sep)
+		w.WriteString(tag.value)
 	}
-	return buf.String()
+	return w.String()
 }