I was looking at this repo and had some time to kill so made some small 

* `serializeTags`: use `strings.Replacer` instead of a regex, use `strings.Builder` and calculate size instead of `bytes.Buffer`
* Don't use `fmt.Sprint()` to concatenate strings: `string1 + string2` is much faster.
* `tcpStatsdSink`: make `sync.Pool` global and shared instead of local to each instance of `tcpStatsdSink` - this is the correct usage since `sync.Pool` is essentially "thread local" (see: [`poolLocalInternal`](https://github.com/golang/go/blob/master/src/sync/pool.go#L56-L61) and [`P`](https://github.com/golang/go/blob/master/src/runtime/proc.go#L22)).
* `tcpStatsdSink.mu`: there is no need for the mutex to be a pointer here - instead give `sync.NewCond()` a reference to it.  The rationale here is that we check this often - so skipping the extra pointer dereference is worth it (additionally, `tcpStatsdSink.flushCond` does not need to be a mutex).
* `nullSink`: is of type `struct{}` it does not need to be pointer.

Benchmark results:
```
benchmark                               old ns/op     new ns/op     delta
BenchmarkStore_MutexContention-8        251           121           -51.79%
BenchmarkStore_NewCounterWithTags-8     2878          571           -80.16%
BenchmarkSerializeTags-8                2596          481           -81.47%

benchmark                               old allocs     new allocs     delta
BenchmarkStore_MutexContention-8        2              0              -100.00%
BenchmarkStore_NewCounterWithTags-8     32             4              -87.50%
BenchmarkSerializeTags-8                29             3              -89.66%

benchmark                               old bytes     new bytes     delta
BenchmarkStore_MutexContention-8        22            2             -90.91%
BenchmarkStore_NewCounterWithTags-8     800           336           -58.00%
BenchmarkSerializeTags-8                688           256           -62.79%
```