@charlievieth 2018-12-19
I can confirm this works on `go1.8.7` the tests fail because `golang.org/x/lint/golint` cannot be built with `go1.7` and `go1.8` due to the addition of aliases, which added the `types.TypeName.IsAlias()` method.