@diegs 2018-12-19
@charlievieth perhaps we should update travis to only support go 1.9+? Would have preferred not to merge this with non-green CI (not sure if CI was green prior to this PR).

Thanks for the improvements though!