@charlievieth 2018-12-19
@diegs Agreed, probably shoulda double checked with you before merging.  Tests broke because of https://github.com/golang/tools/commit/9e2f8b2a0ac2cc9c3237908fdd5c0471b6ab311f.

Second g1.9+ support, see PR #49.